+++ src/macros.rs
@@ -1,17 +1,18 @@
+#[doc(hidden)]
 #[macro_export]
-macro_rules! derive_trait {
+macro_rules! derive_trait_internal {
     ($data:expr, _, $trait:item $(,)*) => {
-        $crate::__rt::derive_trait!($data, None, _, $trait)
+        $crate::__rt::derive_trait_internal!($data, None, _, $trait)
     };
     ($data:expr, ($($path:tt)*), $trait:item $(,)*) => {
-        $crate::__rt::derive_trait!($data, None, ($($path)*), $trait)
+        $crate::__rt::derive_trait_internal!($data, None, ($($path)*), $trait)
     };
     ($data:expr, $super:expr, _, $trait:item $(,)*) => {
         $crate::__rt::parse2($crate::__rt::quote!($trait))
             .map_err($crate::Error::from)
             .and_then(|trait_: $crate::__rt::ItemTrait| {
                 let path = $crate::__rt::path(Some(trait_.ident.clone().into()));
-                $crate::__rt::derive_trait!($data, $super, path, trait_)
+                $crate::__rt::derive_trait_internal!($data, $super, path, trait_)
             })
     };
     ($data:expr, $super:expr, ($($path:tt)*), $trait:item $(,)*) => {
@@ -19,16 +20,38 @@ macro_rules! derive_trait {
             .map_err($crate::Error::from)
             .and_then(|path| {
                 let trait_: $crate::__rt::ItemTrait = $crate::__rt::parse2($crate::__rt::quote!($trait))?;
-                $crate::__rt::derive_trait!($data, $super, path, trait_)
+                $crate::__rt::derive_trait_internal!($data, $super, path, trait_)
             })
     };
-    ($data:expr, $path:expr, $trait:expr $(,)*) => {{
-        $crate::__rt::derive_trait!($data, None, $path, $trait)
-    }};
+    ($data:expr, $path:expr, $trait:expr $(,)*) => {
+        $crate::__rt::derive_trait_internal!($data, None, $path, $trait)
+    };
     ($data:expr, $super:expr, $path:expr, $trait:expr $(,)*) => {{
         let trait_: $crate::__rt::ItemTrait = $trait;
-        $data.impl_trait_with_capacity(trait_.items.len(), $path, $super, trait_).map($crate::build)
+        $data.impl_trait_with_capacity(trait_.items.len(), $path, $super, trait_).map($crate::__rt::build_item)
     }};
+}
+
+#[macro_export]
+macro_rules! derive_trait {
+    ($data:expr, _, $trait:item $(,)*) => {
+        $crate::__rt::derive_trait_internal!($data, _, $trait).map($crate::__rt::ToTokens::into_token_stream)
+    };
+    ($data:expr, ($($path:tt)*), $trait:item $(,)*) => {
+        $crate::__rt::derive_trait_internal!($data, ($($path)*), $trait).map($crate::__rt::ToTokens::into_token_stream)
+    };
+    ($data:expr, $super:expr, _, $trait:item $(,)*) => {
+        $crate::__rt::derive_trait_internal!($data, $super, _, $trait).map($crate::__rt::ToTokens::into_token_stream)
+    };
+    ($data:expr, $super:expr, ($($path:tt)*), $trait:item $(,)*) => {
+        $crate::__rt::derive_trait_internal!($data, $super, ($($path)*), $trait).map($crate::__rt::ToTokens::into_token_stream)
+    };
+    ($data:expr, $path:expr, $trait:expr $(,)*) => {
+        $crate::__rt::derive_trait_internal!($data, $path, $trait).map($crate::__rt::ToTokens::into_token_stream)
+    };
+    ($data:expr, $super:expr, $path:expr, $trait:expr $(,)*) => {
+        $crate::__rt::derive_trait_internal!($data, $super, $path, $trait).map($crate::__rt::ToTokens::into_token_stream)
+    };
 }
 
 /// A macro for to make easy to write `proc_macro_derive` like deriving trait to enum so long as all variants are implemented that trait.